refactor: checkpoint loading now fully functioning including lr scheduler
